fix: Correction d'un probleme pour le dropout
Augmentation de la taille de l'argument d'entrée

+++ include/layers/dense.h
@@ -53,7 +53,7 @@ class Dense
             xt::xarray<float> target,
             float tauxApprentissage);
 
-        void dropout(uint8_t dropRate);
+        void dropout(uint16_t dropRate);
 };
 
 #endif
